requirements.txt
# ============================================================================
# Django related packages
# ============================================================================
Django==1.4.3
django-countries==1.5
django_libs==1.6
django-extensions==1.0.1
django-debug-toolbar==0.9.4
# django-rosetta==0.7.1
-e git+git://github.com/mbi/django-rosetta.git@3ce6ae6e8bab28e4ee56c80350e92ae16ee9bb2a#egg=rosetta
johnny-cache==1.4
urllib3==1.5
django-admin-tools==0.5.1


# ============================================================================
# Other important packages
# ============================================================================
Pillow==1.7.8
South==0.8.1
Markdown==2.2.1
psycopg2==2.4.5
html5lib==0.95
python-memcached==1.48
requests==0.13.5


# ============================================================================
# Multilingual related packages
# ============================================================================
# simple-translation==0.8.6
-e git+git://github.com/mbrochh/simple-translation.git@91c98e0cb1a4150f0460855dea869ddd75aaf25e#egg=simple_translation
cmsplugin-filer-image-translated==0.1

# ============================================================================
# django-cms related packages
# ============================================================================
#django-classy-tags==0.3.4.1
django-cms==2.3.5
django-mptt==0.5.2
django-sekizai==0.6.1
cmsplugin-markdown==0.1
cmsplugin-blog-authors==0.1
cmsplugin_googlemap==0.1.3


# ATTENTION: You MUST NOT use -e git+... for the following modules because they
# are symlinked on the server to
# /virtualenvs/buildee/lib/python2.7/site-packages/modulename in extra_locales
# and when you install the source distribution the symlink will no longer be
# valid
django-people==0.6.4
cmsplugin-blog-categories==0.6
django-document-library==0.6.2
django-multilingual-events==0.6.6
django-tagging-translated==0.1
# END OF ATTENTION


django-localized-names==0.2
django-roadmap==0.1
django-ics-generator==0.1
django-multilingual-orgs==0.3
django-multilingual-initiatives==0.2
django-good-practice-examples==0.2.4
django-generic-positions==0.1

# We need this because it is a dependency of the good-practice-example-app.
django-multilingual-news==0.1.1

# There is a problem with STATIC_URL in this release.
# cmsplugin-blog==1.1.2
-e git+git://github.com/fivethreeo/cmsplugin-blog.git@cd91a4eb066437aa254574759eed51da973314ab#egg=cmsplugin_blog
-e git+git://github.com/bitmazk/cmsplugin-blog-images.git@5030205894e6fc2db1fb42ee0a895055cce3de8f#egg=cmsplugin_blog_images
django-hero-slider==0.2
-e git+git://github.com/bitmazk/cmsplugin-link-extended.git@acec221aef3966d30c89608ef9ad38cd1c1fa7d7#egg=cmsplugin_link_extended
cmsplugin-blog-language-publish==0.1.3

# ============================================================================
# django-filer related packages
# ============================================================================
easy-thumbnails==1.1
django-filer==0.9.3
cmsplugin-filer==0.9.2


# ============================================================================
# Test & development related packages
# ============================================================================
factory_boy==1.2.0
fabric==1.4.4
flake8==1.7.0
watchdog==0.6.0
nose==1.2.1
django-nose==1.1
coverage==3.6
django-coverage==1.2.2
